requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
sqlalchemy>=1.4.23
pymysql>=1.0.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
fastapi-mail>=1.2.0
jinja2>=3.0.1
python-dotenv>=0.19.0
redis>=4.0.0